Pipeline: drop the `base` concept
Each pipeline is now self-contained without references to another.
However, as the final stage in a pipeline is saved to the content
store, we are able to reuse it if one pipeline is the prefix of
another.

Note that if two pipelines merely share a prefix, that does not
help. One pipeline must be a proper prefix of the other.

Signed-off-by: Tom Gundersen <teg@jklm.no>