test: test the in-tree version of our tool
Make sure we test the version of osbuild in the current checkout,
rather than the system instance.

Also default to using in-place directories for the object store
and output images. Using a tmpfs does not scale, especially on
CI infrastructue with limited memory.

The behavior can still be overridden by the environment variable,
as before, only the default changes.

Signed-off-by: Tom Gundersen <teg@jklm.no>