Tree compare: Add experimental python script for mounting qcow2
This commit introduces proof-of-concept script for mounting qcow2
images. It's already working on our CI.

The big decision was if we should use qemu-nbd or libguestfs. Both
solutions has been tested and qemu-nbd has won.

Libguestfs is a really nice and easy-to-use project but it had
sever performance issues on CI (about 8 times slower than qemu-nbd).
As we expect to do deep testing of a resulting image it's nice to have
something faster not to slow down our test running times.
Therefore qemu-nbd was chosen.