tests: Add tree difference tests
This commit introduces a whole new kind of tests: Tree difference test.
The test firstly builds two trees. Then, they are compared using the
tree-diff utility. Finally, the diff is checked against a set of rules.

The tree to be compared can be both output of last stage in pipeline or
output of assembler. In the first case the tree is found inside osbuild
store. It the second case the resulting image from an assembler is
mounted. Both current assemblers (tar and qcow2) are supported.