buildroot: configure nspawn's stdio setup, if needed
With version 242 systemd-nspawn(5) gained the --console argument to
control how stdin, stdout, stderr is set up; with this change the
behavior of systemd-nspawn also changed. Before, only when nspawn
detected that stdio were ttys it would setup stdio as (pseudo-)
terminal. After that change, it will do so if the console mode is
not 'pipe', which needs to be explicitly specified, because the
default, i.e. if stdio is not a tty and nothing is specified, is
'read-only'. This specifically also means that no input is passed
as stdin to the container. But some stages, e.g. dnf, require data
to be delivered via stdin and thus this change breaks those stages.
Revert back to the old behavior of systemd by passing '--console=
pipe'. This means osbuild depends on a systemd >= 242 now.