Add support for building a UEFI bootable disk image
This adds a new stage, org.osbuild.uefi-bls, which copies configuration
files to the UEFI directories and configures the image so that UEFI will
use the bls entries.

It also adds a new assembler, org.osbuild.uefi, which creates a UEFI
disk image that will be booted using the grub2 blscfg command.

Included is a f30 sample that demonstrates building a UEFI bootable
raw image.